File structure update
