--- frontend/src/assets/pages/home.jsx
@@ -1,11 +1,0 @@
-import styles from '../styles/home.module.css';
-
-const Home = () => {
-    return (
-        <div>
-            <h1>Home Page</h1>
-        </div>
-    )
-}
-
-export default Home;
