+++ frontend/src/main.jsx
@@ -1,11 +1,11 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import "./index.css";
 
-import AppRouter from './assets/utils/router.jsx'
+import AppRouter from "./utils/router.jsx";
 
-createRoot(document.getElementById('root')).render(
+createRoot(document.getElementById("root")).render(
   <StrictMode>
     <AppRouter />
-  </StrictMode>,
-)
+  </StrictMode>
+);
