+++ frontend/src/pages/game.jsx
@@ -1,0 +1,11 @@
+import styles from "../styles/game.module.css";
+
+const Game = () => {
+  return (
+    <div>
+      <h1>Game Page</h1>
+    </div>
+  );
+};
+
+export default Game;
