+++ frontend/src/pages/home.jsx
@@ -1,0 +1,11 @@
+import styles from "../styles/home.module.css";
+
+const Home = () => {
+  return (
+    <div>
+      <h1>Home Page</h1>
+    </div>
+  );
+};
+
+export default Home;
